change age to integer

+++ forms.py
@@ -1,10 +1,10 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, FloatField
+from wtforms import StringField, IntegerField
 
 class AddNewPetForm(FlaskForm):
     """Create form for adding new pets"""
     name = StringField("Pet Name")
     species = StringField("Species")
     photo_url = StringField("URL of Photo")
-    age = FloatField("Age")
+    age = IntegerField("Age")
     notes = StringField("Notes")
